Support clang, use size_t, const, fix UD in wait
 - Makefile:  Support clang, add zish.o to OBJS, add warnings
 - Use f(size_t argc, char *argv[argc]) instead of
   f(int argc, char **argv) or f(char **argc, int argc)
 - zish_rawmemchr was claming it's first parameter as const but it
   returned a reference to it as non-const, leading to
   const-incorrectness since the user could be tempted to modify said
   returned pointer, thinking their original const pointer wasn't
   modified though it now is.
 - waitpid() does not always give a valid status, handle that.
 - A bit more verbose perror() messages

+++ zish.c
@@ -58,7 +58,7 @@ static void zish_initialize(void)
         setenv("PS1", "$ ", false);
 
     char *home_path = getenv("HOME");
-    int home_path_size = strlen(home_path);
+    size_t home_path_size = strlen(home_path);
 
     history_full_path = malloc((home_path_size + strlen(history_file) + 2) * sizeof(*history_full_path));
     strcpy(history_full_path, home_path);
@@ -75,7 +75,7 @@ static void zish_initialize(void)
 
     free(config_full_path);
 
-    srand(time(NULL));
+    srand((unsigned int)(time(NULL)));
 
     aliases = calloc(1, sizeof(*aliases));
 }
